requirements.txt
google-genai>=1.0.0
pyyaml>=6.0
jinja2>=3.1
